src/requirements.txt
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0,<3.0.0
gensim>=4.3.0
openai>=1.3.0,<2.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<1.13.0
python-dotenv>=1.0.0
pymysql>=1.1.0
